pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cholerama"
version = "0.1.0"
description = "Competitive multiplayer game of life"
license = {text = "BSD-3-Clause"}
authors = [{name = "Neil Vaytet"}]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Games/Entertainment :: Turn Based Strategy",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent"
]
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pillow",
    "pyqt5 == 5.15.9",
    "pyqtgraph == 0.13.2",
]

[tool.setuptools.packages.find]
where = ["./src"]

[tool.setuptools.package-data]
vendeeglobe = [
    "resources/*",
]
